ENTERPRISE-19; cnange addon groupId to match one registered on Atlassian
