+++ src/aps/log4j.properties
@@ -560,5 +560,5 @@ log4j.additivity.com.atlassian.plugin.remotable.plugin.service.LocalSignedReques
 log4j.logger.com.atlassian.plugin.remotable.host.common.service.http.bigpipe.DefaultBigPipeManager = INFO, console, remoteappssecurity
 log4j.additivity.com.atlassian.plugin.remotable.host.common.service.http.bigpipe.DefaultBigPipeManager = false
 
-log4j.logger.com.codenvy.plugin = DEBUG, console, filelog
-log4j.additivity.com.codenvy.plugin = false
+log4j.logger.com.codenvy.jira = DEBUG, console, filelog
+log4j.additivity.com.codenvy.jira = false
