+++ src/main/java/com/codenvy/jira/customfield/ReviewCustomField.java
@@ -8,7 +8,7 @@
  * Contributors:
  *   Codenvy, S.A. - initial API and implementation
  *******************************************************************************/
-package com.codenvy.plugin.customfield;
+package com.codenvy.jira.customfield;
 
 import com.atlassian.jira.issue.customfields.impl.GenericTextCFType;
 import com.atlassian.jira.issue.customfields.manager.GenericConfigManager;
@@ -17,11 +17,11 @@ import com.atlassian.jira.issue.fields.TextFieldCharacterLengthValidator;
 import com.atlassian.jira.security.JiraAuthenticationContext;
 
 
-public class DevelopCustomField extends GenericTextCFType {
+public class ReviewCustomField extends GenericTextCFType {
 
-    public DevelopCustomField(CustomFieldValuePersister customFieldValuePersister, GenericConfigManager genericConfigManager,
-                              TextFieldCharacterLengthValidator textFieldCharacterLengthValidator,
-                              JiraAuthenticationContext jiraAuthenticationContext) {
+    public ReviewCustomField(CustomFieldValuePersister customFieldValuePersister, GenericConfigManager genericConfigManager,
+                             TextFieldCharacterLengthValidator textFieldCharacterLengthValidator,
+                             JiraAuthenticationContext jiraAuthenticationContext) {
 
         super(customFieldValuePersister, genericConfigManager, textFieldCharacterLengthValidator, jiraAuthenticationContext);
     }
